# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: StoreFlow Backend API
  version: 1.0.0
  description: Unified API for Stores, Projects, Tasks, Milestones, Blockers, Comments, Approvals, and Files.
servers:
  - url: https://api.storeflow.local/api/v1
    description: Local/Example server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id:
      type: string
      description: MongoDB ObjectId
      example: 64b7f1d6f1a2b3c4d5e6f7a8
...
    Project:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/Id' }
        storeId: { $ref: '#/components/schemas/Id' }
        name: { type: string }
        type: { type: string, enum: [setup, renovation, other] }
        status: { type: string, enum: [not_started, in_progress, on_hold, completed, cancelled] }
    Task:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/Id' }
        projectId: { $ref: '#/components/schemas/Id' }
        title: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        priority: { type: string, enum: [low, medium, high, urgent] }
...
paths:
...
  /stores/{storeId}/projects:
    get:
      tags: [Projects]
      summary: List projects in store
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [Projects]
      summary: Create project
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '201': { description: Project created }
  /stores/{storeId}/projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in project
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - name: projectId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
    post:
      tags: [Tasks]
      summary: Create task
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - name: projectId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '201': { description: Task created }
...